Add new serializer for email authentication
I create a new serializer for the email authentication
so it will check if the input data is a email or is a username.
Then it will try to authenticate with that data, if the authentication
operation is sucessful it will return a token in other case it will send
a error message.