Updated functionality create referendum
There was a problem with the creation of the Voting object when creating a referendum because the relationship with the Question type has changed in the model. Previously, this functionality was implemented with a OneToMany relation from the Question type to the Voting type. Now, the relationship is ManyToMany. Closes #104