Updated serializars
I have modified the json of a voting. Now you can see the unlockquestions. You only will be able to see the unlockquestions of the first round of questions, in other words, a voting will have two rounds of questions as maximum, even a option of the second quetions unlocks some new questions, to avoid loops.