Add: Voting Functionality
FIX #88: There is now a votings list available for the votings in which the user is censed. Only opened votings will be available to click on. Others will display Opening date or closed symbol.

FIX #89: There is now a view for the voting booth. The user may select an option and vote.